refactor: QueryLogService 및 관련 테스트 코드 경로 및 로직 수정
1. QueryLogService를 application에서 application.service로 이동한다.
2. UsersRepository로 변경된 구조를 반영하여 관련 메서드 호출부 수정한다.
3. QueryRequest, QueryResponse, TokenCalculator 등 패키지 경로 변경 사항을 적용한다.
4. usersRepository.save 호출 로직 추가로 변경된 도메인 객체 저장한다.
5. 테스트 코드의 Mock 대상 및 관련된 assert 로직을 최신 구조로 업데이트한다.

Resolves #22
